Get LSP light experiment working
