+++ vscode/src/banana/banana.ts
@@ -1,6 +1,4 @@
-export class Banana {
-    constructor(
-        public x: number,
-        public z: number
-    ) {}
+export interface Banana {
+    x: number
+    z: number
 }
