+++ vscode/src/completions/context/lsp-light-graph-cache.ts
@@ -1,18 +1,24 @@
+import path from 'node:path'
+
 import { debounce } from 'lodash'
 import { LRUCache } from 'lru-cache'
 import * as vscode from 'vscode'
+import { URI } from 'vscode-uri'
 
 import { HoverContext } from '@sourcegraph/cody-shared/src/codebase-context/messages'
+import { dedupeWith } from '@sourcegraph/cody-shared/src/common'
 
 import { getGraphContextFromRange } from '../../graph/graph'
-import { ContextSnippet } from '../types'
+import { ContextSnippet, SymbolContextSnippet } from '../types'
 
 import { GraphContextFetcher, supportedLanguageId } from './context-graph'
+import { SectionObserver } from './section-observer'
 import { CustomAbortController, CustomAbortSignal } from './utils'
 
 export class LspLightGraphCache implements vscode.Disposable, GraphContextFetcher {
     private disposables: vscode.Disposable[] = []
     private cache: GraphCache = new GraphCache()
+    private sectionObserver: SectionObserver = SectionObserver.createInstance()
 
     private lastRequestKey: string | null = null
     private abortLastRequest: () => void = () => {}
@@ -20,7 +26,7 @@ export class LspLightGraphCache implements vscode.Disposable, GraphContextFetche
     public static instance: LspLightGraphCache | null = null
     public static createInstance(): LspLightGraphCache {
         if (this.instance) {
-            throw new Error('SectionObserver has already been initialized')
+            throw new Error('LspLightGraphCache has already been initialized')
         }
         this.instance = new LspLightGraphCache()
         return this.instance
@@ -52,19 +58,13 @@ export class LspLightGraphCache implements vscode.Disposable, GraphContextFetche
         const prevLine = Math.max(position.line - 1, 0)
         const currentLine = position.line
 
-        const start = performance.now()
-
-        const [prevLineContext, currentLineContext] = await Promise.all([
-            null,
-            // this.getLspContextForLine(document, prevLine, 0, abortController.signal),
+        const [prevLineContext, currentLineContext, sectionHistory] = await Promise.all([
+            this.getLspContextForLine(document, prevLine, 0, abortController.signal),
             this.getLspContextForLine(document, currentLine, 1, abortController.signal),
+            this.sectionObserver.getSectionHistory(document, position, contextRange),
         ])
 
-        console.log({
-            duration: performance.now() - start,
-            prevLineContext,
-            currentLineContext,
-        })
+        const sectionGraphContext = [...prevLineContext, ...currentLineContext]
 
         if (maxChars === 0) {
             // This is likely just a preloading request, so we don't need to prepare the actual
@@ -72,7 +72,46 @@ export class LspLightGraphCache implements vscode.Disposable, GraphContextFetche
             return []
         }
 
-        return []
+        let usedContextChars = 0
+        const context: ContextSnippet[] = []
+
+        function overlapsContextRange(uri: vscode.Uri, range?: { startLine: number; endLine: number }): boolean {
+            if (!contextRange || !range || uri.toString() !== document.uri.toString()) {
+                return false
+            }
+
+            return contextRange.start.line <= range.startLine && contextRange.end.line >= range.endLine
+        }
+
+        // Allocate up to 40% of the maxChars budget to inlining previous section unless we have no
+        // graph context
+        const maxCharsForPreviousSections = sectionGraphContext ? maxChars * 0.4 : maxChars
+        for (const historyContext of sectionHistory) {
+            if (usedContextChars + historyContext.content.length > maxCharsForPreviousSections) {
+                // We use continue here to test potentially smaller context snippets that might
+                // still fit inside the budget
+                continue
+            }
+            usedContextChars += historyContext.content.length
+            context.push(historyContext)
+        }
+
+        if (sectionGraphContext) {
+            const preciseContexts = hoverContextsToSnippets(
+                sectionGraphContext.filter(context => !overlapsContextRange(URI.parse(context.uri), context.range))
+            )
+            for (const preciseContext of preciseContexts) {
+                if (usedContextChars + preciseContext.content.length > maxChars) {
+                    // We use continue here to test potentially smaller context snippets that might
+                    // still fit inside the budget
+                    continue
+                }
+                usedContextChars += preciseContext.content.length
+                context.push(preciseContext)
+            }
+        }
+
+        return context
     }
 
     private getLspContextForLine(
@@ -153,5 +192,20 @@ class GraphCache {
 
     public delete(key: GraphCacheParams): void {
         this.cache.delete(this.toCacheKey(key))
+    }
+}
+
+function hoverContextsToSnippets(contexts: HoverContext[]): SymbolContextSnippet[] {
+    return dedupeWith(contexts.map(hoverContextToSnippets), context =>
+        [context.symbol, context.fileName, context.content].join('\n')
+    )
+}
+
+function hoverContextToSnippets(context: HoverContext): SymbolContextSnippet {
+    return {
+        fileName: path.normalize(vscode.workspace.asRelativePath(URI.parse(context.uri).fsPath)),
+        symbol: context.symbolName,
+        sourceSymbolAndRelationship: undefined,
+        content: context.content.join('\n').trim(),
     }
 }
