+++ vscode/src/graph/graph.ts
@@ -73,19 +73,17 @@ export const getGraphContextFromRange = async (
 ): Promise<HoverContext[]> => {
     const uri = document.uri
     contentMap = contentMap.size === 0 ? new Map([[uri.fsPath, document.getText().split('\n')]]) : contentMap
-    const selections = [{ uri, range }]
+    const locations = [{ uri, range }]
 
     const label = 'getGraphContextFromRange'
     performance.mark(label)
 
     // Find the candidate identifiers to request definitions for in the selection
-    const requestCandidates = gatherDefinitionRequestCandidates(selections, contentMap)
-
-    console.log({ requestCandidates })
+    const requestCandidates = gatherDefinitionRequestCandidates(locations, contentMap)
 
     // Extract hover (symbol, definition, type def, impl) text related to all of the request
     // candidates
-    const resolvedHoverText = await gatherHoverText(contentMap, requestCandidates, abortSignal)
+    const resolvedHoverText = await gatherHoverText(contentMap, requestCandidates, abortSignal, recursionLimit > 0)
 
     const contexts = resolvedHoverText.flatMap(hoverContextFromResolvedHoverText)
 
@@ -93,28 +91,31 @@ export const getGraphContextFromRange = async (
     performance.mark(label)
 
     if (recursionLimit > 0) {
-        console.log({ contexts: [...contexts] })
-        console.log('RECURS', recursionLimit)
-        const locations = dedupeWith(
-            contexts.flatMap(context => {
-                if (!context.range) {
-                    return []
-                }
-                const range = new vscode.Range(
-                    context.range.startLine,
-                    context.range.startCharacter,
-                    context.range.endLine,
-                    context.range.endCharacter
-                )
-                return [new vscode.Location(URI.parse(context.uri), range)]
-            }),
-            item => locationKeyFn(item)
+        const locationsForRecursion = dedupeWith(resolvedHoverText.map(r => r.definition).filter(isDefined), item =>
+            locationKeyFn(item)
         )
 
+        // Open all documents needed for the recursion and ensure contentMap is up to date
+        const uniqueUris = dedupeWith(locationsForRecursion, l => l.uri.toString()).map(l => l.uri)
+        const documents = await Promise.all(
+            uniqueUris.map(async (uri: vscode.Uri) => {
+                const document = await vscode.workspace.openTextDocument(uri)
+                if (!contentMap.has(uri.fsPath)) {
+                    contentMap.set(uri.fsPath, document.getText().split('\n'))
+                }
+                return [uri.toString(), document] as const
+            })
+        )
+        const documentsMap: Map<string, vscode.TextDocument> = new Map(documents)
+
         const recursiveContexts = await Promise.all(
-            locations.map(async location => {
-                const document = await vscode.workspace.openTextDocument(location.uri)
-                return getGraphContextFromRange(document, range, abortSignal, recursionLimit - 1, contentMap)
+            locationsForRecursion.map(async location => {
+                const document = documentsMap.get(location.uri.toString())
+                if (!document) {
+                    throw new Error('Document not found')
+                }
+
+                return getGraphContextFromRange(document, location.range, abortSignal, recursionLimit - 1, contentMap)
             })
         )
 
@@ -130,7 +131,7 @@ interface Selection {
 }
 
 /**
- * Return the definitions of symbols that occur within the given selection ranges. If a selection
+ * Return the d efinitions of symbols that occur within the given selection ranges. If a selection
  * has a defined range, we will cull the symbols to those referenced in intersecting document symbol
  * ranges.
  */
@@ -145,22 +146,22 @@ const getGraphContextFromSelection = async (
     // Get the document symbols in the current file and extract their definition range
     const definitionSelections = await extractRelevantDocumentSymbolRanges(selections)
 
-    // Find the candidate identifiers to request definitions for in the selection
+    // Find the candida te identifiers to request definitions for in the selection
     const requestCandidates = gatherDefinitionRequestCandidates(definitionSelections, contentMap)
 
-    // Extract identifiers from the relevant document symbol ranges and request their definitions
+    // Extract identifi ers from the relevant document symbol ranges and request their definitions
     const definitionMatches = await gatherDefinitions(definitionSelections, requestCandidates)
 
-    // NOTE: Before asking for data about a document it must be opened in the workspace. This forces
-    // a resolution so that the following queries that require the document context will not fail with
-    // an unknown document.
+    // NOTE: Before ask ing for data about a document it must be opened in the workspace. This forces
+    // a resolution so  that the following queries that require the document context will not fail with
+    // an unknown docum ent.
 
     await updateContentMap(
         contentMap,
         definitionMatches.map(({ definitionLocations }) => definitionLocations.map(({ uri }) => uri)).flat()
     )
 
-    // Resolve, extract, and deduplicate the symbol and location match pairs from the definition matches
+    // Resolve, extract , and deduplicate the symbol and location match pairs from the definition matches
     const matches = dedupeWith(
         definitionMatches
             .map(({ definitionLocations, typeDefinitionLocations, implementationLocations, ...rest }) =>
@@ -172,7 +173,7 @@ const getGraphContextFromSelection = async (
 
     // TODO - see if we can remove fields of types we've also captured?
 
-    // Extract definition text from our matches
+    // Extract definiti on text from our matches
     const contexts = await extractDefinitionContexts(matches, contentMap)
 
     logDebug('GraphContext:snippetsRetrieved', `Retrieved ${contexts.length} context snippets`)
@@ -202,13 +203,13 @@ const getGraphContextFromSelection = async (
 }
 
 /**
- * Open each URI referenced by a definition match in the current workspace, and make the document
+ * Open each UR I referenced by a definition match in the current workspace, and make the document
  * content retrievable by filepath by adding it to the shared content map.
  */
 const updateContentMap = async (contentMap: Map<string, string[]>, locations: vscode.Uri[]): Promise<void> => {
     const unseenDefinitionUris = dedupeWith(locations, 'fsPath').filter(uri => !contentMap.has(uri.fsPath))
 
-    // Remove ultra-common type definitions that are probably already known by the LLM
+    // Remove ultra-com mon type definitions that are probably already known by the LLM
     const filteredUnseenDefinitionUris = unseenDefinitionUris.filter(uri => !isCommonImport(uri))
 
     const newContentMap = new Map(
@@ -224,7 +225,7 @@ const updateContentMap = async (contentMap: Map<string, string[]>, locations: vs
 }
 
 /**
- * Get the document symbols in files indicated by the given selections and extract the symbol
+ * Get the docu ment symbols in files indicated by the given selections and extract the symbol
  * ranges. This will give us indication of where either the user selection and cursor is located or
  * the range of a relevant definition we've fetched in a previous iteration, which we assume to be
  * the most relevant code to the current question.
@@ -254,9 +255,9 @@ export const extractRelevantDocumentSymbolRanges = async (
             continue
         }
 
-        // Filter the document symbol ranges to just those whose range intersects the selection.
-        // If no selection exists (if we have an undefined in the ranges list), keep all symbols,
-        // we'll utilize all document symbol ranges.
+        // Filter the docum ent symbol ranges to just those whose range intersects the selection.
+        // If no selection  exists (if we have an undefined in the ranges list), keep all symbols,
+        // we'll utilize al l document symbol ranges.
         const definedRanges = ranges.filter(isDefined)
         combinedRanges.push(
             ...(definedRanges.length < ranges.length
@@ -300,7 +301,7 @@ const goKeywords = new Set([
     'type',
     'var',
 
-    // common variables, types we don't need to follow
+    // common variables , types we don't need to follow
     'ctx',
     'Context',
     'err',
@@ -411,18 +412,18 @@ const pythonKeywords = new Set([
 export const commonKeywords = new Set([...goKeywords, ...typescriptKeywords, ...pythonKeywords])
 
 const commonImportPaths = new Set([
-    // The TS lib folder contains the TS standard library and all of ECMAScript.
+    // The TS lib folde r contains the TS standard library and all of ECMAScript.
     'node_modules/typescript/lib',
     // The node library contains the standard node library.
     'node_modules/@types/node',
-    // All CSS properties as TS types.
+    // All CSS properti es as TS types.
     'node_modules/csstype',
-    // Common React type definitions.
+    // Common React typ e definitions.
     'node_modules/@types/react/',
     'node_modules/@types/prop-types',
     'node_modules/next/',
 
-    // Go stdlib installation (covers Brew installs at a minimum)
+    // Go stdlib instal lation (covers Brew installs at a minimum)
     'libexec/src/',
 
     // Python stdlib
@@ -446,17 +447,16 @@ interface Request {
 }
 
 /**
- * Search the given ranges identifier definitions matching an a common identifier pattern and filter out
+ * Search the g iven ranges identifier definitions matching an a common identifier pattern and filter out
  * common keywords.
  */
 export const gatherDefinitionRequestCandidates = (
-    selections: Selection[],
+    locations: vscode.Location[],
     contentMap: Map<string, string[]>
 ): Request[] => {
     const requestCandidates: Request[] = []
 
-    for (const selection of selections) {
-        const { uri, range } = selection
+    for (const { uri, range } of locations) {
         const lines = contentMap.get(uri.fsPath)
         if (!range || !lines) {
             continue
@@ -464,10 +464,10 @@ export const gatherDefinitionRequestCandidates = (
 
         for (const { start, end } of [range]) {
             for (const [lineIndex, line] of lines.slice(start.line, end.line + 1).entries()) {
-                // NOTE: pretty hacky - strip out C-style line comments and find everything
-                // that might look like it could be an identifier. If we end up running a
-                // VSCode provider over this cursor position and it's not a symbol we can
-                // use, we'll just get back an empty location list.
+                // NOTE: pretty hac ky - strip out C-style line comments and find everything
+                // that might look  like it could be an identifier. If we end up running a
+                // VSCode provider  over this cursor position and it's not a symbol we can
+                // use, we'll just  get back an empty location list.
                 const identifierMatches = line.replace(/\/\/.*$/, '').matchAll(identifierPattern)
 
                 for (const match of identifierMatches) {
@@ -505,7 +505,7 @@ interface ResolvedSymbolDefinitionMatches {
 }
 
 /**
- * Query each of the candidate requests for definitions which are resolved in parallel before return.
+ * Query each o f the candidate requests for definitions which are resolved in parallel before return.
  */
 export const gatherDefinitions = async (
     selections: Selection[],
@@ -516,13 +516,13 @@ export const gatherDefinitions = async (
     getImplementations: typeof defaultGetImplementations = defaultGetImplementations
 ): Promise<ResolvedSymbolDefinitionMatches[]> => {
     // Construct a list of symbol and definition location pairs by querying the LSP server
-    // with all identifiers (heuristically chosen via regex) in the relevant code ranges.
+    // with all identif iers (heuristically chosen via regex) in the relevant code ranges.
     const definitionMatches: SymbolDefinitionMatches[] = []
 
-    // NOTE: deduplicating here will save duplicate queries that are _likely_ to point to the
+    // NOTE: deduplicat ing here will save duplicate queries that are _likely_ to point to the
     // same definition, but we may be culling aggressively here for some edge cases. I don't
-    // currently think that these are likely to be make-or-break a quality response on any
-    // significant segment of real world questions, though.
+    // currently think  that these are likely to be make-or-break a quality response on any
+    // significant segm ent of real world questions, though.
     for (const { symbolName, uri, position } of dedupeWith(requests, 'symbolName')) {
         definitionMatches.push({
             symbolName,
@@ -533,7 +533,7 @@ export const gatherDefinitions = async (
         })
     }
 
-    // Resolve all in-flight promises in parallel
+    // Resolve all in-f light promises in parallel
     const resolvedDefinitionMatches = await Promise.all(
         definitionMatches.map(
             async ({ symbolName, hover, definitionLocations, typeDefinitionLocations, implementationLocations }) => ({
@@ -548,7 +548,7 @@ export const gatherDefinitions = async (
 
     return (
         resolvedDefinitionMatches
-            // Remove definition ranges that exist within one of the input definition selections
+            // Remove definitio n ranges that exist within one of the input definition selections
             // These are locals and don't give us any additional information in the context window.
             .map(({ definitionLocations, ...rest }) => ({
                 definitionLocations: definitionLocations.filter(
@@ -563,7 +563,7 @@ export const gatherDefinitions = async (
                 ),
                 ...rest,
             }))
-            // Remove empty locations
+            // Remove empty loc ations
             .filter(
                 ({ definitionLocations, typeDefinitionLocations, implementationLocations }) =>
                     definitionLocations.length + typeDefinitionLocations.length + implementationLocations.length !== 0
@@ -571,20 +571,12 @@ export const gatherDefinitions = async (
     )
 }
 
-interface UnresolvedHoverText {
-    symbolName: string
-    symbolLocation: vscode.Location
-    definitionsPromise: Thenable<vscode.Location[]>
-    typeDefinitionsPromise: Thenable<vscode.Location[]>
-    implementationsPromise: Thenable<vscode.Location[]>
-}
-
 interface ResolvedHoverText {
     symbolName: string
     symbolLocation: vscode.Location
-    definition?: ResolvedHoverElement
+    symbol: ResolvedHoverElement
+    definition?: vscode.Location
     typeDefinition?: ResolvedHoverElement
-    implementations?: ResolvedHoverElement[]
 }
 
 interface ResolvedHoverElement {
@@ -593,19 +585,17 @@ interface ResolvedHoverElement {
     hover: vscode.Hover[]
 }
 
-const hoverToStrings = (h: vscode.Hover[]): string[] =>
-    h
-        .flatMap(h => h.contents.map(c => (typeof c === 'string' ? c : c.value)))
+const hoverToStrings = (hover: vscode.Hover[]): string[] =>
+    hover
+        .flatMap(hover => hover.contents.map(c => (typeof c === 'string' ? c : c.value)))
         .map(extractMarkdownCodeBlock)
         .map(s => s.trim())
         .filter(s => s !== '')
 
 const hoverContextFromResolvedHoverText = (t: ResolvedHoverText): HoverContext[] =>
     [
-        hoverContextFromElement(t.definition, 'definition'),
+        hoverContextFromElement(t.symbol, 'symbol'),
         hoverContextFromElement(t.typeDefinition, 'typeDefinition', t.symbolName),
-        ...(t.implementations?.map(e => hoverContextFromElement(e, 'implementation', t.typeDefinition?.symbolName)) ??
-            []),
     ].filter(isDefined)
 
 const hoverContextFromElement = (
@@ -619,10 +609,8 @@ const hoverContextFromElement = (
 
     let content = hoverToStrings(element.hover)
 
-    // Filter out common hover texts that do not provide additional value
-    content = content.filter(
-        c => c.trim() !== `interface ${element.symbolName}` && c.trim() !== `class ${element.symbolName}`
-    )
+    // Filter out commo n hover texts that do not provide additional value
+    content = content.filter(content => !isTypeAlias(element.symbolName, content))
 
     if (content.length === 0) {
         return undefined
@@ -663,138 +651,67 @@ function extractMarkdownCodeBlock(string: string): string {
 }
 
 /**
- * Query each of the candidate requests for hover texts which are resolved in parallel before return.
+ * Query each o f the candidate requests for hover texts which are resolved in parallel before return.
  */
 export const gatherHoverText = async (
     contentMap: Map<string, string[]>,
     requests: Request[],
     abortSignal?: CustomAbortSignal,
+    includeDefinition: boolean = false,
     getHover: typeof defaultGetHover = defaultGetHover,
     getDefinitions: typeof defaultGetDefinitions = defaultGetDefinitions,
-    getTypeDefinitions: typeof defaultGetTypeDefinitions = defaultGetTypeDefinitions,
-    getImplementations: typeof defaultGetImplementations = defaultGetImplementations
+    getTypeDefinitions: typeof defaultGetTypeDefinitions = defaultGetTypeDefinitions
 ): Promise<ResolvedHoverText[]> => {
-    // Construct a list of symbol and definition location pairs by querying the LSP server
-    // with all identifiers (heuristically chosen via regex) in the relevant code ranges.
-    const hoverMatches: UnresolvedHoverText[] = []
+    const symbolLocations = requests.map(({ symbolName, uri, position }) => ({
+        symbolName,
+        symbolLocation: new vscode.Location(uri, position),
+    }))
 
-    // NOTE: deduplicating here will save duplicate queries that are _likely_ to point to the
-    // same definition, but we may be culling aggressively here for some edge cases. I don't
-    // currently think that these are likely to be make-or-break a quality response on any
-    // significant segment of real world questions, though.
-
-    for (const { uri, symbolName, position } of dedupeWith(requests, 'symbolName')) {
-        console.log('hovering over', symbolName, uri.toString(), JSON.stringify(position))
-        hoverMatches.push({
-            symbolName,
-            symbolLocation: new vscode.Location(uri, position),
-            definitionsPromise: limiter(() => getDefinitions(uri, position), abortSignal),
-            typeDefinitionsPromise: limiter(() => getTypeDefinitions(uri, position), abortSignal),
-            implementationsPromise: limiter(() => getImplementations(uri, position), abortSignal),
-        })
-    }
-
-    // Resolve the definition/type definition/implementations queries above in parallel and extract
-    // and deduplicate the locations. We're going to request hover text from each of these next.
-    const locationsForHover = dedupeWith(
-        (
-            await Promise.all(
-                hoverMatches.map(
-                    async ({ symbolLocation, definitionsPromise, typeDefinitionsPromise, implementationsPromise }) => [
-                        symbolLocation,
-                        ...(await definitionsPromise),
-                        ...(await typeDefinitionsPromise),
-                        ...(await implementationsPromise),
-                    ]
-                )
-            )
-        ).flat(),
-        locationKeyFn
-    )
-
-    console.log(locationsForHover)
-
-    // NOTE: Before asking for data about a document it must be opened in the workspace. This forces
-    // a resolution so that the following queries that require the document context will not fail with
-    // an unknown document.
-    await updateContentMap(contentMap, dedupeWith(locationsForHover.map(l => l.uri).flat(), 'fsPath'))
-
-    // Request hover for every (deduplicated) location range we got from def/type def/impl queries
-    const hoverMap = new Map(
-        // Dedupe the locations, we don't want to hover the same range twice
-        dedupeWith(
-            locationsForHover.filter(l => !isCommonImport(l.uri)),
-            item => locationKeyFn(item)
-        ).map(
-            l =>
-                [locationKeyFn(l), limiter(() => getHover(l.uri, l.range.start), abortSignal)] as [
-                    string,
-                    Thenable<vscode.Hover[]>,
-                ]
-        )
-    )
-    const resolvedHoverMap = await unwrapThenableMap(hoverMap)
+    const dedupedSymbolLocations = dedupeWith(symbolLocations, s => locationKeyFn(s.symbolLocation))
 
     return Promise.all(
-        hoverMatches.map(
-            async ({
+        dedupedSymbolLocations.map(async ({ symbolName, symbolLocation }) => {
+            const hoverPromise = limiter(() => getHover(symbolLocation.uri, symbolLocation.range.start), abortSignal)
+            const definitionsPromise = includeDefinition
+                ? limiter(() => getDefinitions(symbolLocation.uri, symbolLocation.range.start), abortSignal)
+                : undefined
+
+            const hover = await hoverPromise
+
+            const contents = hoverToStrings(hover)
+            const isAlias = !!contents.find(content => isTypeAlias(symbolName, content))
+            const typeDefinitionPromise = isAlias
+                ? limiter(() => getTypeDefinitions(symbolLocation.uri, symbolLocation.range.start), abortSignal)
+                      .then(definitions => {
+                          return lastElement(definitions)
+                      })
+                      .then(async definition => {
+                          if (!definition) {
+                              return undefined
+                          }
+
+                          const symbolName = extractRangeFromDocument(contentMap, definition.uri, definition.range)
+
+                          return {
+                              symbolName,
+                              location: definition,
+                              hover: await limiter(() => getHover(definition.uri, definition.range.start), abortSignal),
+                          }
+                      })
+                : undefined
+
+            return {
                 symbolName,
                 symbolLocation,
-                definitionsPromise,
-                typeDefinitionsPromise,
-                implementationsPromise,
-            }) => {
-                let definitionObj: ResolvedHoverElement | undefined
-                const definitionLocation = (await definitionsPromise).pop()
-                if (definitionLocation) {
-                    const symbolName = extractRangeFromDocument(
-                        contentMap,
-                        definitionLocation.uri,
-                        definitionLocation.range
-                    )
-
-                    definitionObj = {
-                        symbolName,
-                        location: definitionLocation,
-                        hover: resolvedHoverMap.get(locationKeyFn(definitionLocation)) || [],
-                    }
-                }
-
-                let typeDefinitionObj: ResolvedHoverElement | undefined
-                const typeDefinitionLocation = (await typeDefinitionsPromise).pop()
-                if (typeDefinitionLocation) {
-                    const symbolName = extractRangeFromDocument(
-                        contentMap,
-                        typeDefinitionLocation.uri,
-                        typeDefinitionLocation.range
-                    )
-
-                    typeDefinitionObj = {
-                        symbolName,
-                        location: typeDefinitionLocation,
-                        hover: resolvedHoverMap.get(locationKeyFn(typeDefinitionLocation)) || [],
-                    }
-                }
-
-                let implementationObjs: ResolvedHoverElement[] | undefined
-                const implementationsLocations = await implementationsPromise
-                if (implementationsLocations.length > 0) {
-                    implementationObjs = implementationsLocations.map(location => ({
-                        symbolName: extractRangeFromDocument(contentMap, location.uri, location.range),
-                        location,
-                        hover: resolvedHoverMap.get(locationKeyFn(location)) || [],
-                    }))
-                }
-
-                return {
+                symbol: {
                     symbolName,
-                    symbolLocation,
-                    definition: definitionObj,
-                    typeDefinition: typeDefinitionObj,
-                    implementations: implementationObjs,
-                }
+                    location: symbolLocation,
+                    hover: await hoverPromise,
+                },
+                definition: definitionsPromise ? lastElement(await definitionsPromise) : undefined,
+                typeDefinition: typeDefinitionPromise ? await typeDefinitionPromise : undefined,
             }
-        )
+        })
     )
 }
 
@@ -804,14 +721,14 @@ const extractRangeFromDocument = (contentMap: Map<string, string[]>, uri: vscode
         return ''
     }
 
-    // Trim off lines outside of the range
-    // NOTE: inclusive upper bound
+    // Trim off lines o utside of the range
+    // NOTE: inclusive  upper bound
     const extractedLines = content.slice(range.start.line, range.end.line + 1)
     if (extractedLines.length === 0) {
         return ''
     }
 
-    // Trim off characters outside of the range
+    // Trim off charact ers outside of the range
     const n = extractedLines.length - 1
     extractedLines[n] = extractedLines[n].slice(0, range.end.character)
     extractedLines[0] = extractedLines[0].slice(range.start.character)
@@ -819,7 +736,7 @@ const extractRangeFromDocument = (contentMap: Map<string, string[]>, uri: vscode
     return extractedLines.join('\n')
 }
 /**
- * For each match, extract the definition text from the given map of file contents. The given content map
+ * For each mat ch, extract the definition text from the given map of file contents. The given content map
  * is expected to hold the contents of the file indicated by the definition's location URI, and the file
  * is assumed to be open in the current VSCode workspace. Matches without such an entry are skipped.
  */
@@ -828,22 +745,22 @@ export const extractDefinitionContexts = async (
     contentMap: Map<string, string[]>,
     getDocumentSymbolRanges: typeof defaultGetDocumentSymbolRanges = defaultGetDocumentSymbolRanges
 ): Promise<PreciseContext[]> => {
-    // Retrieve document symbols for each of the open documents, which we will use to extract the relevant
-    // definition "bounds" given the range of the definition symbol (which is contained within the range).
+    // Retrieve documen t symbols for each of the open documents, which we will use to extract the relevant
+    // definition "boun ds" given the range of the definition symbol (which is contained within the range).
     const documentSymbolsMap = new Map(
         [...contentMap.keys()]
             .filter(fsPath => matches.some(({ location }) => location.uri.fsPath === fsPath))
             .map(fsPath => [fsPath, getDocumentSymbolRanges(vscode.Uri.file(fsPath))])
     )
 
-    // NOTE: In order to make sure the loop below is unblocked we'll also force resolve the entirety
-    // of the folding range requests. That way we don't have a situation where the first iteration of
-    // the loop is waiting on the last promise to be resolved in the set.
+    // NOTE: In order t o make sure the loop below is unblocked we'll also force resolve the entirety
+    // of the folding r ange requests. That way we don't have a situation where the first iteration of
+    // the loop is wait ing on the last promise to be resolved in the set.
     await Promise.all([...documentSymbolsMap.values()])
 
     // Piece everything together. For each matching definition, extract the relevant lines given the
-    // containing document's content and folding range result. Downstream consumers of this function
-    // are expected to filter and re-rank these results as needed for their specific use case.
+    // containing docum ent's content and folding range result. Downstream consumers of this function
+    // are expected to  filter and re-rank these results as needed for their specific use case.
 
     const contexts: PreciseContext[] = []
     for (const { symbolName, hover, location } of matches) {
@@ -853,7 +770,7 @@ export const extractDefinitionContexts = async (
 
         if (contentPromise && documentSymbolsPromises) {
             const content = contentPromise
-            const documentSymbols = await documentSymbolsPromises // NOTE: already resolved
+            const documentSymbols = await documentSymbolsPromises // NOTE: already re solved
 
             const definitionSnippets = extractSnippets(content, documentSymbols, [range])
 
@@ -880,7 +797,7 @@ export const extractDefinitionContexts = async (
 }
 
 /**
- * Shim for default LSP executeDocumentSymbolProvider call. Can be mocked for testing.
+ * Shim for def ault LSP executeDocumentSymbolProvider call. Can be mocked for testing.
  */
 export const defaultGetDocumentSymbolRanges = async (uri: URI): Promise<vscode.Range[]> =>
     vscode.commands
@@ -896,13 +813,13 @@ export const defaultGetDocumentSymbolRanges = async (uri: URI): Promise<vscode.R
         })
 
 /**
- * Shim for default LSP executeHoverPRovider call. Can be mocked for testing.
+ * Shim for def ault LSP executeHoverPRovider call. Can be mocked for testing.
  */
 const defaultGetHover = async (uri: URI, position: vscode.Position): Promise<vscode.Hover[]> =>
     vscode.commands.executeCommand<vscode.Hover[]>('vscode.executeHoverProvider', uri, position)
 
 /**
- * Shim for default LSP executeDefinitionProvider call. Can be mocked for testing.
+ * Shim for def ault LSP executeDefinitionProvider call. Can be mocked for testing.
  */
 const defaultGetDefinitions = async (uri: URI, position: vscode.Position): Promise<vscode.Location[]> =>
     vscode.commands
@@ -910,7 +827,7 @@ const defaultGetDefinitions = async (uri: URI, position: vscode.Position): Promi
         .then(locations => locations.flatMap(extractLocation))
 
 /**
- * Shim for default LSP executeTypeDefinitionProvider call. Can be mocked for testing.
+ * Shim for def ault LSP executeTypeDefinitionProvider call. Can be mocked for testing.
  */
 const defaultGetTypeDefinitions = async (uri: URI, position: vscode.Position): Promise<vscode.Location[]> =>
     vscode.commands
@@ -921,12 +838,12 @@ const defaultGetTypeDefinitions = async (uri: URI, position: vscode.Position): P
         )
         .then(locations => locations.flatMap(extractLocation))
         // Type definitions are not always well-defined for things like functions. In these cases
-        // we'd like to fall back to a regular definition result which gives us the same class and
-        // quality of information.
+        // we'd like to fal l back to a regular definition result which gives us the same class and
+        // quality of infor mation.
         .then(locations => (locations.length > 0 ? locations : defaultGetDefinitions(uri, position)))
 
 /**
- * Shim for default LSP executeImplementationProvider call. Can be mocked for testing.
+ * Shim for def ault LSP executeImplementationProvider call. Can be mocked for testing.
  */
 const defaultGetImplementations = async (uri: URI, position: vscode.Position): Promise<vscode.Location[]> =>
     vscode.commands
@@ -938,7 +855,7 @@ const defaultGetImplementations = async (uri: URI, position: vscode.Position): P
         .then(locations => locations.flatMap(extractLocation))
 
 /**
- * Extract the definition range from the given symbol information or document symbol.
+ * Extract the  definition range from the given symbol information or document symbol.
  */
 const extractSymbolRange = (d: vscode.SymbolInformation | vscode.DocumentSymbol): vscode.Range =>
     isDocumentSymbol(d) ? d.range : d.location.range
@@ -947,7 +864,7 @@ const isDocumentSymbol = (s: vscode.SymbolInformation | vscode.DocumentSymbol): 
     (s as vscode.DocumentSymbol).range !== undefined
 
 /**
- * Convert the given location or location link into a location.
+ * Convert the  given location or location link into a location.
  */
 const extractLocation = (l: vscode.Location | vscode.LocationLink): vscode.Location =>
     isLocationLink(l) ? new vscode.Location(l.targetUri, l.targetRange) : l
@@ -956,25 +873,25 @@ const isLocationLink = (l: vscode.Location | vscode.LocationLink): l is vscode.L
     (l as vscode.LocationLink).targetUri !== undefined
 
 /**
- * Extract the content outlined by symbol ranges that intersect one of the target ranges.
+ * Extract the  content outlined by symbol ranges that intersect one of the target ranges.
  */
 const extractSnippets = (lines: string[], symbolRanges: vscode.Range[], targetRanges: vscode.Range[]): string[] => {
     const intersectingRanges = symbolRanges.filter(fr =>
         targetRanges.some(r => fr.start.line <= r.start.line && r.end.line <= fr.end.line)
     )
 
-    // NOTE: inclusive upper bound
+    // NOTE: inclusive  upper bound
     return intersectingRanges.map(fr => lines.slice(fr.start.line, fr.end.line + 1).join('\n'))
 }
 
 /**
- * Returns a key unique to a given location for use with `dedupeWith`.
+ * Returns a ke y unique to a given location for use with `dedupeWith`.
  */
 export const locationKeyFn = (location: vscode.Location): string =>
     `${location.uri?.fsPath}?L${location.range.start.line}:${location.range.start.character}`
 
 /**
- * Convert a mapping from K -> Thenable<V> to a map of K -> V.
+ * Convert a ma pping from K -> Thenable<V> to a map of K -> V.
  */
 const unwrapThenableMap = async <K, V>(map: Map<K, Thenable<V>>): Promise<Map<K, V>> => {
     const resolved = new Map<K, V>()
@@ -982,4 +899,16 @@ const unwrapThenableMap = async <K, V>(map: Map<K, Thenable<V>>): Promise<Map<K,
         resolved.set(k, await v)
     }
     return resolved
+}
+
+// TODO: Fix this f or Go and Python
+function isTypeAlias(symbolName: string, hover: string): boolean {
+    const trimmed = hover.trim()
+    return (
+        trimmed === `interface ${symbolName}` || trimmed === `class ${symbolName}` || trimmed === `type ${symbolName}`
+    )
+}
+
+function lastElement<T>(array: T[]): T | undefined {
+    return array.at(-1)
 }
